Refactor: 향수 검색 로직 중앙화 및 레이어 분리
