+++ src/components/domains/brandDetail/perfumes/PageClient.tsx
@@ -1,25 +1,16 @@
 "use client";
 
-import { useEffect, useMemo } from "react";
+import { useMemo } from "react";
 import SortDropdown from "@/components/commons/dropdown/SortDropdown";
 import {
   usePerfumeNoteFilter,
   usePerfumeAccordFilter,
-} from "@/lib/hooks/query/useFilters";
-import { useFilterStore } from "@/lib/stores/useFilterStore";
-import { useTotalStore } from "@/lib/stores/useCountStore";
+} from "@/lib/hooks/query/useFilterQuery";
 import { SearchHeader } from "../../../commons/perfumeList/search";
 import { PerfumeSection } from "../../../commons/perfumeList/section/PerfumeSection";
-import {
-  fetchPerfumes,
-  getUniquePerfumes,
-} from "@/components/commons/perfumeList/perfumes.helpers";
-import { InfiniteData, useInfiniteQuery } from "@tanstack/react-query";
+import { getUniquePerfumes } from "@/components/commons/perfumeList/perfumes.helpers";
 import { useInfiniteScrollTrigger } from "@/lib/hooks/useInfiniteScrollTrigger";
-import { queryKeys } from "@/lib/utils/queryKeys";
-import { PaginatedSearchResponse } from "@/lib/hono/schemas/search.schema";
-
-type BrandDetailQueryKey = ReturnType<typeof queryKeys.perfume.brandDetail>;
+import { useInfinitePerfumes } from "@/lib/hooks/useInfinitePerfumes";
 
 export const PageClient = ({ brandName }: { brandName: string }) => {
   const { data: notes } = usePerfumeNoteFilter();
@@ -28,34 +19,18 @@ export const PageClient = ({ brandName }: { brandName: string }) => {
   const memoizedNotes = useMemo(() => notes ?? [], [notes]);
   const memoizedAccords = useMemo(() => accords ?? [], [accords]);
 
-  const filters = useFilterStore((state) => state.filters);
-  const setCount = useTotalStore((state) => state.setTotalCount);
-
-  const { data, isLoading, fetchNextPage, hasNextPage, isFetchingNextPage } =
-    useInfiniteQuery<
-      PaginatedSearchResponse,
-      Error,
-      InfiniteData<PaginatedSearchResponse>,
-      BrandDetailQueryKey,
-      string | null
-    >({
-      queryKey: queryKeys.perfume.brandDetail(brandName),
-      queryFn: ({ pageParam }) => fetchPerfumes(pageParam, brandName, filters),
-      getNextPageParam: (lastPage) => lastPage.nextCursor,
-      initialPageParam: null,
-    });
-
-  useEffect(() => {
-    if (data?.pages[0]?.totalCount) {
-      setCount(data.pages[0].totalCount);
-    }
-  }, [data, setCount]);
+  const {
+    perfumes,
+    isLoading,
+    fetchNextPage,
+    hasNextPage,
+    isFetchingNextPage,
+  } = useInfinitePerfumes(brandName);
 
   // 중복 아이디 제거
   const uniquePerfumes = useMemo(() => {
-    const allPerfumes = data?.pages.flatMap((page) => page.data) ?? [];
-    return getUniquePerfumes(allPerfumes);
-  }, [data]);
+    return getUniquePerfumes(perfumes);
+  }, [perfumes]);
 
   const moreRef = useInfiniteScrollTrigger({
     fetchNextPage,
