+++ src/components/domains/perfumes/PageClient.tsx
@@ -4,19 +4,14 @@ import {
   useBrandFilter,
   usePerfumeNoteFilter,
   usePerfumeAccordFilter,
-} from "@/lib/hooks/query/useFilters";
+} from "@/lib/hooks/query/useFilterQuery";
 import { BrandSection } from "@/components/commons/perfumeList/section/BrandSection";
 import { PerfumeSection } from "@/components/commons/perfumeList/section/PerfumeSection";
 import { SearchHeader } from "@/components/commons/perfumeList/search";
 import SortDropdown from "@/components/commons/dropdown/SortDropdown";
-import { useInfinitePerfumes } from "./hook/useInfinitePerfumes";
+import { useInfinitePerfumes } from "@/lib/hooks/useInfinitePerfumes";
 import { usePerfumeSearchState } from "./hook/usePerfumeSearchState";
 import { useInfiniteScrollTrigger } from "@/lib/hooks/useInfiniteScrollTrigger";
-
-export type BrandName = {
-  en: string;
-  ko: string;
-};
 
 export default function PageClient() {
   const { data: brands } = useBrandFilter();
